+++ src/decorators/validator/common/type.ts
@@ -1,6 +1,7 @@
 enum ParamType {
   VALIDATOR = 'Validator',
   TRANSFORMER = 'Transformer',
+  MAPPING = 'Mapping',
 }
 
 export { ParamType };