+++ src/decorators/validator/common/validateSchema.ts
@@ -1,6 +1,5 @@
-import { validationMetadataKey, Validator } from '../type';
+import { validationMetadataKey, Validator, Transformer, Mappinger } from '../type';
 import { ParamType } from './type';
-import { Transformer } from '../transform/type';
 
 const validateSchema = (target: any, inputData: any) => {
   const neededValidateParams: string[] = Reflect.getOwnMetadata(
@@ -24,6 +23,10 @@ const validateSchema = (target: any, inputData: any) => {
       if (validator.getType === ParamType.TRANSFORMER) {
         const transformedData = (validator as Transformer).process(inputData[param]);
         inputData[param] = transformedData;
+      }
+
+      if (validator.getType === ParamType.MAPPING) {
+        (validator as Mappinger).process(inputData, param);
       }
     });
   });