+++ src/decorators/validator/number/isLess.ts
@@ -3,7 +3,7 @@ import { NameValidator } from '../error/validationError';
 import { validate } from '../method';
 import { required } from '../parameter';
 import { messageMapper, Validator } from '../type';
-import { addNeededValidateParams, addValidatorForParams } from '../../utils/metadata';
+import { addNeededValidateParams, addHandlerForParams } from '../../utils/metadata';
 
 class LessValidator extends BaseValidator implements Validator {
   protected name: NameValidator = NameValidator.IS_LESS;
@@ -29,7 +29,7 @@ function IsLess<T extends Error>(lowerBound: number, errorHandler?: T | string) 
     validator.setErrorHandler(errorHandler);
     validator.setLowerBound(lowerBound);
 
-    addValidatorForParams(target, propertyKey, validator);
+    addHandlerForParams(target, propertyKey, validator);
   };
 }